splitting up my C file
