+++ src/quibble_verses.c
@@ -1,0 +1,243 @@
+/*-------------verses---------------------------------------------------------//
+
+ Purpose: A quibble verse is a user-submitted function fragement to be compiled
+          later
+
+//----------------------------------------------------------------------------*/
+
+#include "../include/quibble_program.h"
+
+bool qb_is_verse(char *verse, int offset){
+    char substr[8] = "__verse";
+    for (int i = 0; i < 7; ++i){
+        if (verse[i+offset] != substr[i]){
+            return false;
+        }
+    }
+
+    return true;
+}
+
+quibble_verse qb_parse_verse(char *verse){
+    int verse_size = strlen(verse);
+    int offset = 0;
+    bool qbinlined = qb_is_inlined(verse);
+    if (qbinlined){
+        offset += 22;
+    }
+
+    if (qb_is_verse(verse, offset)){
+        offset += 7;
+    }
+    else {
+        fprintf(stderr, "Quibble verses must be configured with `__verse`!\n");
+        exit(1);
+    }
+
+    quibble_verse final_verse;
+
+    int config_start = qb_find_next_char(verse, offset, '(')+1;
+    int config_end = qb_find_matching_char(verse, verse_size, config_start-1, '(', ')');
+
+
+    if (config_end-config_start > 0){
+        char *config =
+            (char *)calloc((config_end-config_start)+1, sizeof(char));
+
+        for (int i = 0; i < (config_end-config_start); ++i){
+            config[i] = verse[config_start+i];
+        }
+
+        final_verse.num_args = qb_find_number_of_args(config);
+        final_verse.args =
+            qb_parse_args(config, final_verse.num_args);
+
+        final_verse.num_kwargs = qb_find_number_of_kwargs(config);
+        final_verse.kwargs =
+            qb_parse_kwargs(config, final_verse.num_kwargs);
+        free(config);
+    }
+    else {
+        final_verse.args = NULL;
+        final_verse.num_args = 0;
+        final_verse.kwargs = NULL;
+        final_verse.num_kwargs = 0;
+    }
+
+    int body_start = qb_find_next_char(verse, offset, '{')+1;
+    int body_end =
+        qb_find_matching_char(verse, verse_size, body_start-1, '{', '}');
+
+    if (body_end-body_start > 0){
+        char *body = (char *)calloc((body_end-body_start)+1, sizeof(char));
+        for (int i = 0; i < (body_end-body_start); ++i){
+            body[i] = verse[body_start+i];
+        }
+
+        if (qbinlined){
+            qb_preprocess_content(body);
+        }
+        final_verse.body = body;
+    }
+    else{
+        final_verse.body = NULL;
+    }
+    final_verse.name = qb_strip_spaces(verse, offset, config_start-2);
+    final_verse.echo = false;
+
+    return final_verse;
+}
+
+// Configures prologues of existing verses
+// The variadic function takes triples after the initial verse and number of
+// kwargs being modified:
+//     1. char * name of variable
+//     2. char * type of value
+//     3. auto value to be used
+void qb_configure_verse_variadic(quibble_verse *qv, int n, va_list args){
+    for (int i = 0; i < n; ++i){
+        char *curr_variable = va_arg(args, char *);
+        int kwarg_index = qb_find_kwarg_index(qv->kwargs,
+                                              qv->num_kwargs,
+                                              curr_variable);
+        free(qv->kwargs[kwarg_index].value);
+
+        char *curr_type = va_arg(args, char *);
+
+        if (strcmp(curr_type, "int") == 0   ||
+            strcmp(curr_type, "i") == 0     ||
+            strcmp(curr_type, "d") == 0){
+
+            int arg = va_arg(args, int);
+            qv->kwargs[kwarg_index].value = qb_int_to_string(arg);
+        }
+        else if (strcmp(curr_type, "float") == 0  ||
+                 strcmp(curr_type, "double") == 0 ||
+                 strcmp(curr_type, "f") == 0      ||
+                 strcmp(curr_type, "lf") == 0){
+            double arg = va_arg(args, double);
+            qv->kwargs[kwarg_index].value = qb_float_to_string(arg);
+        }
+        else if (strcmp(curr_type, "quibble_array") == 0 ||
+                 strcmp(curr_type, "qa") == 0){
+            qv->kwargs[kwarg_index].value = 
+                qb_array_to_string(va_arg(args, quibble_array));
+        }
+        else if (strcmp(curr_type, "quibble_variable") == 0 ||
+                 strcmp(curr_type, "qv") == 0){
+            qv->kwargs[kwarg_index].value = 
+                qb_variable_to_string(va_arg(args, quibble_variable));
+        }
+        else{
+            fprintf(stderr,
+                    "Type %s not supported as quibble input!\n",
+                    curr_type);
+            exit(1);
+        }
+    }
+
+}
+
+void qb_configure_verse(quibble_verse *qv, int n, ...){
+    va_list args;
+    va_start(args, n);
+
+    qb_configure_verse_variadic(qv, n, args);
+ 
+    va_end(args);
+}
+
+// An echo is a verse with the same body, but different prologue
+quibble_verse qb_echo_verse(quibble_verse qv, int n, ...){
+    quibble_verse final_qv;
+    final_qv.echo = true;
+    final_qv.body = qv.body;
+    final_qv.name = qv.name;
+    final_qv.num_kwargs = qv.num_kwargs;
+    if (qv.num_kwargs > 0){
+        final_qv.kwargs =
+            (quibble_kwarg *)malloc(sizeof(quibble_kwarg)*qv.num_kwargs);
+    }
+    else {
+        final_qv.kwargs = NULL;
+    }
+
+    for (int i = 0; i < qv.num_kwargs; ++i){
+        int vlength = strlen(qv.kwargs[i].variable)+1;
+        final_qv.kwargs[i].variable = (char *)malloc(sizeof(char)*vlength);
+
+        for (int j = 0; j < vlength; ++j){
+            final_qv.kwargs[i].variable[j] = qv.kwargs[i].variable[j];
+        }
+
+        vlength = strlen(qv.kwargs[i].value)+1;
+        final_qv.kwargs[i].value = (char *)malloc(sizeof(char)*vlength);
+        for (int j = 0; j < vlength; ++j){
+            final_qv.kwargs[i].value[j] = qv.kwargs[i].value[j];
+        }
+    }
+
+    va_list args;
+    va_start(args, n);
+ 
+    qb_configure_verse_variadic(&final_qv, n, args);
+ 
+    va_end(args);
+
+    return final_qv;
+}
+
+char *qb_expand_verse(quibble_program qp, char* verse_name, char *config){
+
+    int idx = qb_find_verse_index(qp, verse_name);
+    if (qp.verse_list[idx].body == NULL){
+        return NULL;
+    }
+    char *tmp_prologue = qb_create_prologue(config, verse_name,
+        qp.verse_list[idx].args, qp.verse_list[idx].num_args,
+        qp.verse_list[idx].kwargs, qp.verse_list[idx].num_kwargs);
+
+    char *tmp_body = (char *)calloc(MAX_META_SIZE, sizeof(char));
+    strcat(tmp_body, tmp_prologue);
+    strcat(tmp_body, qp.verse_list[idx].body);
+
+    int len = strlen(tmp_body);
+    char *final_body = (char *)calloc(len+5, sizeof(char));
+
+    final_body[0] = '{';
+    final_body[1] = '\n';
+
+    for (int i = 2; i < len+2; ++i){
+        final_body[i] = tmp_body[i-2];
+    }
+
+    final_body[len+2] = '}';
+    final_body[len+3] = '\n';
+    
+    free(tmp_body);
+    free(tmp_prologue);
+    return final_body;
+    
+}
+
+void qb_free_verse(quibble_verse qv){
+    if (qv.echo == false){
+        free(qv.body);
+        free(qv.name);
+    }
+    qb_free_kwarg_array(qv.kwargs, qv.num_kwargs);
+    qb_free_arg_array(qv.args, qv.num_args);
+}
+
+void qb_print_verse(quibble_verse qv){
+    printf("Quibble Verse:\nName: %s\nBody:\n%s\n", qv.name, qv.body);
+    printf("Args...\n");
+    for (int i = 0; i < qv.num_args; ++i){
+        qb_print_arg(qv.args[i]);
+    }
+    printf("Kwargs...\n");
+    for (int i = 0; i < qv.num_kwargs; ++i){
+        qb_print_kwarg(qv.kwargs[i]);
+    }
+    printf("\n");
+}
